feat: LMS updates — B2C YouTube catalog, thumbnails API, security/CORS, dashboard, CourseBrowse UI
- Backend: B2C JSON catalog, media sync, CourseService/LessonService thumbnail resolution
- Root API endpoint + permit GET / and OPTIONS for CORS; start-local.ps1 frees port 8081
- Frontend: Navbar, CourseBrowse, progress API client, session expiry handling
- Tests: H2 dialect for integration tests; docs and schema additions

Made-with: Cursor

+++ backend/src/main/java/com/learnova/dto/course/CourseDto.java
@@ -10,6 +10,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class CourseDto {
+
     private Long id;
     private String title;
     private String description;
@@ -18,7 +19,7 @@ public class CourseDto {
     private String category;
     private String instructorName;
     private Long instructorId;
-    private Long totalLessons;
+    private Integer totalLessons;
     private Long totalDurationSeconds;
     private Boolean enrolled;
 }
